Adding RPATH entries
RPATH needs to include a few additional entries (in the right order) for
in-tree binaries to execute correctly. While its not perfect, the
"build" directories have been added to the end to ensure this is always
the case (this is suboptimal as it embeds build structure info into the
final executable, but it's not harmful due to the order).